refactor(game): move constants into lib/core (T27)
Relocate constants.ts to packages/game/lib/core/ to keep all
core-layer modules colocated. Seven importers updated and the
vitest coverage exclude path updated to match the new location.
